@@ -3,9 +3,8 @@ import { Component, Input } from '@angular/core';
 @Component({
   selector: 'app-svg',
   templateUrl: './svg.component.html',
-  styleUrls: ['./svg.component.scss']
+  styleUrls: ['./svg.component.scss'],
 })
 export class SvgComponent {
-
-  @Input() name : string = ''
+  @Input() name: string = '';
 }
